Add shift_gen option to get_io_combi
WARNING: This breaks if more than 10 generations are present.
This allows to get combinations that are shifted in their generation index by 1.
This is the easiest solution I could come up with to avoid pairing equal
files, if the data and reference lambdas agree.

NOTE: that there is some _tension_ between the cl args and the actually used
args currently.