Add script for doing a chi2 scan of the parameter space.
This is the first part of a series of scripts, that can be used to
make chi2 scans of the parameter space and then differently combining
them to get results from the combined values.
The script scans the paramter space and stores the values into a
pandas.DataFrame, that then gets written into a file. From there other
scripts (that are still to be implemented) will take up and do their
part.