Improve memory footprint of chi2 parameter scan.
Abandoning using a dictionary for each parameter set in favor of a
list reduces the peak memory usage by about a factor 2, while not
being slower.
Replacing the pre-allocated list of lists with a 2D numpy array,
yields about another factor 2 of better memory performance but the
speed is lowered by almost 20 %, so that there is now a switch at a
given size (currently 30 M scan points), where the tradeoff of speed
vs memory is set.  The switching point is set such, that the memory
usage for the list of list case doesn't exceed 7 Gbytes of memory (too
far).

It is not entirely clear to me where the 20 % speed difference come
from. It doesn't seem to be related to the memory layout of the numpy
array as column and row major layout run at the same speed. Some quick
performance checks also didn't reveal anything really obvious to
me. There are less LLC loads and CPU cycles in the list of list
approach, but only in the 4 - 6 % range (also the LLC load misses are
lower for the numpy array version).

Maybe it's the (python) overhead of indexing into the numpy array?